requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
joblib==1.2.0
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.6
